DP-5404 - Use generic BIZ API instead of specialized API. Add heffing_valuta_code and heffing_display. Various other improvements

+++ src/biz/dataset/models.py
@@ -14,7 +14,7 @@ class BIZData(models.Model):
     biz_type = models.CharField(max_length=64, blank=True, null=True)
     heffingsgrondslag = models.CharField(max_length=128, blank=True, null=True)
     website = models.CharField(max_length=128, blank=True, null=True)
-    heffing = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
+    heffing = models.IntegerField(blank=True, null=True)
     bijdrageplichtigen = models.IntegerField(blank=True, null=True)
     verordening = models.CharField(max_length=128, blank=True, null=True)
     geometrie = models.GeometryField(db_column='wkb_geometry', srid=28992, blank=True, null=True)
